WIP: Bug fix in JSON value streaming
